Bruk stsklient?
